A/requirements.txt
pystray
pillow
pyinstaller
